\name{CoCoRaw-internal}
\title{Internal CoCoRaw objects}
\alias{.base}
\alias{.clean.up}
\alias{.current}
\alias{.decompose.models}
\alias{.dispose.of.fitted.values}
\alias{.eh.enter.base.model}
\alias{.enter.double.list}
\alias{.enter.n.interactions}
\alias{.First.lib}
\alias{.fix.edges}
\alias{.names.from.model}
\alias{.omit.test}
\alias{.onAttach}
\alias{.onLoad}
\alias{.plotCoCo}
\alias{.return.complex}
\alias{.set.datastructure}
\alias{.set.ic}
\alias{.set.switch}
\alias{.show.log.lik}
\alias{.split.model.gc}
\alias{.split.name.set}
\alias{.sub.split.names}

%    \alias{CoCoIdClass-class}
%    \alias{CoCoClass-class}
%    \alias{CoCoModelClass-class}
%    \alias{CoCoTestClass-class}
%    
%    \alias{label} % CoCoTestClass-method
%    \alias{testLabel} % CoCoTestClass-method
%    \alias{width} % CoCoTestClass-method
%    \alias{testWidth} % CoCoTestClass-method
%    
%    \alias{dynamic.Graph} % CoCoModelClass-method
%    \alias{dynamic.Graph.Old} % CoCoModelClass-method
%    \alias{modifyModel} % CoCoModelClass-method
%    \alias{subModifyModel} % CoCoModelClass-method
%    \alias{subModifyModel.Old} % CoCoModelClass-method
%    \alias{testEdge} % CoCoModelClass-method
%    \alias{testEdge.Old} % CoCoModelClass-method
%    
%    \alias{makeModel} % CoCoObjects
%    \alias{newCoCoTestObject} % CoCoObjects
%    \alias{recover.coco.model} % CoCoObjects
%    

     \alias{coco.init} % From CoCoCore

%    \alias{coco.resume} % From CoCoCore
%    \alias{coco.start} % From CoCoCore
%    \alias{currentCoCo} % From CoCoObjects
%    \alias{endCoCo} % From CoCoRaw
%    \alias{is.number} % From CoCoCore
%    \alias{make.cococg} % From CoCoCore
%    \alias{make.coco} % From CoCoCore
%    \alias{makeCurrentCoCo} % From CoCoObjects
%    \alias{make.model} % From CoCoCore
%    \alias{my.assign} % From CoCoCore
%    \alias{quitCoCo} % From CoCoRaw

\description{Internal CoCoRaw and imported objects.}

\details{Some of these are not to be called by the user.}

\keyword{internal}
